// src/app/modules/master/organization-level/Constants/OrgLevels.enum.ts
export enum OrgLevel {
    One = "OrgLevel1",
    Two = "OrgLevel2",
    Three = "OrgLevel3",
    Four = "OrgLevel4"
}

export enum OrgLevelCode {
    One = "OrgLevel1Code",
    Two = "OrgLevel2Code",
    Three = "OrgLevel3Code",
    Four = "OrgLevel4Code"
}

export enum Status{
    Active = 'Active',
    Inactive = 'Inactive',
    All = 'All',
    Activate = 'Activate',
    activate = 'activate',
    Deactivate = 'Deactivate',
    Disabled = 'Disabled',
    Status = 'Status'
}

export enum NavigationUrls {
    Org1Add = 'xrm/master/organization-level1/add-edit',
    Org1Edit = '/xrm/master/organization-level1/add-edit/',
    Org1View = '/xrm/master/organization-level1/view/',
    Org1List = '/xrm/master/organization-level1',
    Org2Add = 'xrm/master/organization-level2/add-edit',
    Org2Edit = '/xrm/master/organization-level2/add-edit/',
    Org2View = '/xrm/master/organization-level2/view/',
    Org2List = '/xrm/master/organization-level2',
    Org3Add = 'xrm/master/organization-level3/add-edit',
    Org3Edit = '/xrm/master/organization-level3/add-edit/',
    Org3View = '/xrm/master/organization-level3/view/',
    Org3List = '/xrm/master/organization-level3',
    Org4Add = 'xrm/master/organization-level4/add-edit',
    Org4Edit = '/xrm/master/organization-level4/add-edit/',
    Org4View = '/xrm/master/organization-level4/view/',
    Org4List = '/xrm/master/organization-level4'
}

export enum ToastMessages{
    OrgLevel1HasBeenSavedSuccessfully = 'OrgLevel1HasBeenSavedSuccessfully',
    OrgLevel1HasBeenDeactivated = 'OrgLevel1HasBeenDeactivated',
    OrgLevel1HasBeenActivated = 'OrgLevel1HasBeenActivated',
    MassOrgLevel1HasBeenDeactivated = 'MassOrgLevel1HasBeenDeactivated',
    MassOrgLevel1HasBeenActivated = 'MassOrgLevel1HasBeenActivated',
    OrgLevel2HasBeenSavedSuccessfully = 'OrgLevel2HasBeenSavedSuccessfully',
    OrgLevel2HasBeenDeactivated = 'OrgLevel2HasBeenDeactivated',
    OrgLevel2HasBeenActivated = 'OrgLevel2HasBeenActivated',
    MassOrgLevel2HasBeenDeactivated = 'MassOrgLevel2HasBeenDeactivated',
    MassOrgLevel2HasBeenActivated = 'MassOrgLevel2HasBeenActivated',
    OrgLevel3HasBeenSavedSuccessfully = 'OrgLevel3HasBeenSavedSuccessfully',
    OrgLevel3HasBeenDeactivated = 'OrgLevel3HasBeenDeactivated',
    OrgLevel3HasBeenActivated = 'OrgLevel3HasBeenActivated',
    MassOrgLevel3HasBeenDeactivated = 'MassOrgLevel3HasBeenDeactivated',
    MassOrgLevel3HasBeenActivated = 'MassOrgLevel3HasBeenActivated',
    OrgLevel4HasBeenSavedSuccessfully = 'OrgLevel4HasBeenSavedSuccessfully',
    OrgLevel4HasBeenDeactivated = 'OrgLevel4HasBeenDeactivated',
    OrgLevel4HasBeenActivated = 'OrgLevel4HasBeenActivated',
    MassOrgLevel4HasBeenDeactivated = 'MassOrgLevel4HasBeenDeactivated',
    MassOrgLevel4HasBeenActivated = 'MassOrgLevel4HasBeenActivated',
    EntityHasBeenActivatedSuccessfully = "EntityHasBeenActivatedSuccessfully",
    EntityHasBeenDeactivatedSuccessfully = "EntityHasBeenDeactivatedSuccessfully",
    Somethingwentwrong = 'Somethingwentwrong',
    EnitityAlreadyExists = 'EnitityAlreadyExists',
}

export enum ValidationMessages {
    PleaseSelectData = 'PleaseSelectData',
    PleaseEnterData = 'PleaseEnterData'
}

export enum RequiredStrings {
    Sector = 'Sector',
    EmptyString = '',
    BasicTitle = 'basic-title',
    Text = 'Text',
    Value = 'Value',
}

export enum ListApis{
    Org1GetAll = 'org1-all-paging',
    Org1AdvanceSearch = 'org1-advance-search-drpdwn',
    Org2GetAll = 'org2-all-paging',
    Org2AdvanceSearch = 'org2-advance-search-drpdwn',
    Org3GetAll = 'org3-all-paging',
    Org3AdvanceSearch = 'org3-advance-search-drpdwn',
    Org4GetAll = 'org4-all-paging',
    Org4AdvanceSearch = 'org4-advance-search-drpdwn'
}